risolto bug creazione conservazione

+++ Database/snake_case_sql_dump/procedure.sql
@@ -280,7 +280,7 @@ DELIMITER ;
 DELIMITER ;;
 CREATE PROCEDURE `create_conservazione`(IN icodice VARCHAR(2), IN inome VARCHAR(20),
 
-    IN iprob_estinzione INT, IN idescrizione TEXT)
+    IN iprob_estinzione decimal(6,5), IN idescrizione TEXT)
 BEGIN
 
         INSERT INTO conservazione (codice, nome, prob_estinzione, descrizione)
